You only need to implement BaseSocialApp to have your own detector for a certain app. Each app would have a unique name, and implemented by get_name function Main.py is now only for demonstration of how to use the API. Will work on it soon to extend to cmd too
PhoneNumber now uses dict to store whether a phone uses an app or not. The key for dict will be the app's name and result will be one of the constants in AppUsageEnum

Added SocialDetector. This class will be the main interface of the program. Right now it has two functions; to add social app and to process the result of detection into phone numbers. Will work on extending that in the near future

Added class Utils. It will have static methods that will be used in various places in the program. Right now it only has processing of command arguments into PhoneNumber objects

Signed-off-by: Yazeed Almuqwishi <g--56@hotmail.com>